feat: camera shots

+++ gl_lib/gl.h
@@ -17,5 +17,5 @@ void glTriangle(Vector3, Vector3, Vector3);
 void glMap(string);
 void glTexture(string);
 void glDraw();
-void glLoadMatrix(float*, int*, float*);
+void glLoadMatrix(float*, float*, float*);
 void glEye(float*, float*, float*);
